Исправления и добавление нового файла
