+++ js/data.js
@@ -1,42 +1,13 @@
 import {getRandomInt,getRandomArrayEl} from './util.js';
-
-const SIMILLAR_OBJECTS_AMOUNT = 25;
-
-const COMMENTATOR_NAMES = [
-  'Илья',
-  'Вика',
-  'Фёдор',
-  'Настя',
-  'Алексей',
-  'Жанна',
-];
-
-const COMMENTS = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const DESCPRIPTION = [
-  'Картинка',
-  'Красивая картнка',
-  'Очень красивая картинка',
-  'Классно',
-  'Хорошо',
-];
-
+import {SIMILLAR_OBJECTS_AMOUNT, COMMENTATOR_NAMES, COMMENTS, DESCPRIPTION} from './const.js'
 
 
 const getAvatarPath = function () {
   return `img/avatar-${getRandomInt(1, 6)}.svg`
 };
 
-const idArray = [];
-
 const getUniqId = function () {
+  const idArray = [];
   const rand = getRandomInt(0, 10000);
   if (!idArray.includes(rand)) {
     idArray.push(rand);
@@ -84,3 +55,4 @@ const generateObjectsArray = function () {
 
 generateObjectsArray();
 
+export {generateObjectsArray};
